requirements.txt
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.4
Cython==0.25.2
defusedxml==0.5.0
Django==1.11.2
django-allauth==0.32.0
django-appconf==1.0.2
django-compressor==2.1.1
django-htmlmin==0.10.0
django-localflavor==1.5.1
django-resized==0.3.5
django-uuslug==1.1.8
djangorestframework==3.6.3
html5lib==0.999999999
idna==2.5
numpy==1.13.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.1.1
pystan==2.16.0.0
python-slugify==1.2.4
python3-openid==3.1.0
pytz==2017.2
rcssmin==1.0.6
requests==2.18.1
requests-oauthlib==0.8.0
rjsmin==1.0.12
six==1.10.0
Unidecode==0.4.21
urllib3==1.21.1
webencodings==0.5.1
